Move stream_as_generator() to utils
